+++ app/models/payload.rb
@@ -3,5 +3,11 @@ module TrafficSpy
     # validates :identifier, presence: true, uniqueness: true
     belongs_to :user
     belongs_to :resolution
+
+
+    def self.url_popularity
+      group(:url).count
+    end
+
   end
 end
