+++ test/controllers/user_receives_correct_response_when_submitting_payload_test.rb
@@ -2,6 +2,7 @@ require_relative '../test_helper'
 # require 'json'
 
 class PayloadTest < TrafficTest
+  include PayloadPrep
 
   def register_user(identifier)
     post '/sources', {rootUrl: 'http://turing.io', identifier: identifier}
