+++ test/models/payload_parsing_test.rb
@@ -1,11 +1,8 @@
 require_relative '../test_helper'
 
 class PayloadParserTest < TrafficTest
-
-  def register_user
-    parser_setup({"rootUrl"=>"http://jumpstartlab.com", "identifier"=>"jumpstartlab"}).parsing_validating
-  end
-
+  include PayloadPrep
+  
   def payload_params
     {"payload"=>
       "{\"url\":\"http://jumpstartlab.com/blog\",\"requestedAt\":\"2013-02-16 21:38:28 -0700\",\"respondedIn\":37,\"referredBy\":\"http://jumpstartlab.com\",\"requestType\":\"GET\",\"parameters\":[],\"eventName\":\"socialLogin\",\"userAgent\":\"Mozilla/5.0 (Macintosh; Intel Mac OS X 10_8_2) AppleWebKit/537.17 (KHTML, like Gecko) Chrome/24.0.1309.0 Safari/537.17\",\"resolutionWidth\":\"1920\",\"resolutionHeight\":\"1280\",\"ip\":\"63.29.38.211\"}",
@@ -14,9 +11,7 @@ class PayloadParserTest < TrafficTest
      "id"=>"jumpstartlab"}
   end
 
-  def parser_setup(params)
-    @parser = TrafficSpy::RegistrationParser.new(params)
-  end
+
 
   def test_payload_parser_parses_properly
     payload_parser = TrafficSpy::PayloadParser.new(payload_params)
