+++ test/models/payload_test.rb
@@ -1,0 +1,11 @@
+# require_relative '../test_helper'
+#
+# class PayloadTest < FeatureTest
+#   include PayloadPrep
+#
+#
+#   def test_url_popularity
+# 
+#   end
+#
+# end
